Move noir related js tests to src
