+++ packages/noir-ethereum-api/src/noir/encode.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { decodeField, decodeHexAddress, encodeAddress, encodeBytes32 } from '../src/noir/encode.js';
+import { decodeField, decodeHexAddress, encodeAddress, encodeBytes32 } from './encode.js';
 
 describe('encodeBytes32', () => {
   it('zero', () => {
